hf_space/requirements.txt
ultralytics>=8.0
streamlit>=1.30
pandas
numpy
Pillow
opencv-python-headless
